🔥 create one file module
