--- vue-todo/src/store/getters.js
@@ -1,3 +1,0 @@
-export const storedTodoItems = (state) => {
-    return state.todoItems;
-}
